x64 portability (commit 1)
git-svn-id: http://miranda.googlecode.com/svn/trunk@9175 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/src/modules/database/database.cpp
@@ -65,7 +65,7 @@ int getProfilePath(TCHAR * buf, size_t cch)
 // returns 1 if *.dat spec is matched
 int isValidProfileName(const TCHAR *name)
 {
-	int len = _tcslen(name) - 4;
+	size_t len = _tcslen(name) - 4;
 	return len > 0 && _tcsicmp(&name[len], _T(".dat")) == 0;
 }
